Show add info container if newuser
